- Common.Types: rendu la méthode WeakList.GetCopyOfList publique, pour permettre à d'autres d'en profiter. - Common.Support: adapté AbstractEntity (DataContextId et déplacé GetNakedEntityName vers EntityInfo<T>). - DataLayer: rendu DataContext.UniqueId 64-bit pour éviter d'éventuelles mauvaises surprises avec des applications qui tournent pendant des mois sans interruption. - DataLayer: DataContextPool est nettement plus performant quand on cherche si une entité appartient à un DataContext, dans la mesure où l'on peut utiliser DataContextId. - DataLayer: DataContextPool est prêt pour une utilisation de multiples instances.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2010.1@15860 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Cresus.DataLayer/Saver/PersistenceJobConverter.cs
@@ -72,7 +72,7 @@ namespace Epsitec.Cresus.DataLayer.Saver
 		/// <returns>The converted sequence of <see cref="AbstractSynchronizationJob"/>.</returns>
 		private IEnumerable<AbstractSynchronizationJob> ConvertHelper(DeletePersistenceJob job)
 		{
-			int dataContextId = this.DataContext.UniqueId;
+			long dataContextId = this.DataContext.UniqueId;
 			EntityKey entityKey = this.DataContext.GetNormalizedEntityKey (job.Entity).Value;
 
 			yield return new DeleteSynchronizationJob (dataContextId, entityKey);
@@ -108,7 +108,7 @@ namespace Epsitec.Cresus.DataLayer.Saver
 		{
 			if (job.JobType == PersistenceJobType.Update)
 			{
-				int dataContextId = this.DataContext.UniqueId;
+				long dataContextId = this.DataContext.UniqueId;
 				EntityKey entityKey = this.DataContext.GetNormalizedEntityKey (job.Entity).Value;
 
 				foreach (var update in job.GetFieldIdsWithValues ())
@@ -151,7 +151,7 @@ namespace Epsitec.Cresus.DataLayer.Saver
 		{
 			if (job.JobType == PersistenceJobType.Update)
 			{
-				int dataContextId = this.DataContext.UniqueId;
+				long dataContextId = this.DataContext.UniqueId;
 				EntityKey sourceKey = this.DataContext.GetNormalizedEntityKey (job.Entity).Value;
 				Druid fieldId = job.FieldId;
 
@@ -200,7 +200,7 @@ namespace Epsitec.Cresus.DataLayer.Saver
 		{
 			if (job.JobType == PersistenceJobType.Update)
 			{
-				int dataContextId = this.DataContext.UniqueId;
+				long dataContextId = this.DataContext.UniqueId;
 				EntityKey sourceKey = this.DataContext.GetNormalizedEntityKey (job.Entity).Value;
 				Druid fieldId = job.FieldId;
